api/requirements.txt
openai>=1.12.0
psycopg2-binary>=2.9.9
pywebpush>=1.14.0
